+++ R/DroughtData-package.R
@@ -1,5 +1,8 @@
 #' @keywords internal
 #' @importFrom tibble tibble
+#' @importFrom rlang .data
+#' @importFrom rlang :=
+#' @importFrom magrittr %>%
 "_PACKAGE"
 
 # The following block is used by usethis to automatically manage